--- Lesson8/src/main/java/HomeWork.java
@@ -1,41 +1,0 @@
-public class HomeWork {
-    //1) задача!
-    /*
-    Необходимо написать военкомат.
-    + 1) Создаем класс военкомат model.MilitaryOffice
-    2) Создать в классе model.MilitaryOffice конструктор, который принимает model.PersonRegistry
-    3) model.PersonRegistry содержит список людей (model.Person), которые могут быть призваны в армию
-    + 4) Создать класс model.Person, который будет содержать следующие поля имя, возраст, пол, адрес(страна, город)
-    Военкомат должен уметь:
-    a) вывести имена всех людей годных к военной службе на текущий момент времени, которые есть в model.PersonRegistry
-    T.е нужно написать метод в классе model.MilitaryOffice, который будет отфильтровывать годных к службе призывников
-    (годными считать мущин от 18 до 27 лет, проверяем пол и возраст)
-    б) вывести количество годных призывников в городе Минске.
-    в) вывести количество призывников от 25 до 27 лет
-    г) вывести количество призывников у которых имя Александр.
-    5) Создать класс Task1.Main, наполнить model.PersonRegistry людьми, по желанию можно создавать людей через консоль
-    */
-
-    //2) задача!
-    /*
- Напишите структуру классов и продемонстрируйте работу
- - Машина имеет двигатель, бензобак (реализуйте класс для каждой сущности).
- - Чтобы поехать, машину необходимо завести, т.е. включить двигатель.
- - Реализуйте методы включения машины, который в свою очередь включает её двигатель.
- - Реализуйте метод езды на машине (например просто печатаем на консоль, что машина поехала)
- - Если машина не заведена, ехать она не может.
- - Машину можно заглушить.
- - После каждой поездки считаем, что машина прошла фиксированное расстояние.
- - Реализовать возможность посмотреть, какое расстояние машина прошла за все время.
- - Чтобы создать машину обязательно нужно иметь двигатель и бензобак.
- - Марка машины, год выпуска, пройденное расстояние - не обязательны при создании машины и могут быть выставлены потом.
- (не обязательно задавать в конструкторе)
- - После создания поменять двигатель машине нельзя.
- - Чтобы машина завелась, у неё должно быть топливо в бензобаке, если топлива нет, машина не может завестись.
- - Машину можно дозаправить, можно проверить сколько топлива осталось.
- - Реализуйте пару полей для двигателя и бензобака, например: тип двигателя, общий объем бензобака,
- сколько бензина сейчас и т.д.
-     */
-
-
-}
