+++ Lesson8/src/main/java/by/tms/task1/model/PersonRegistry.java
@@ -1,7 +1,10 @@
 //model.PersonRegistry содержит список людей (model.Person), которые могут быть призваны в армию
 
-package Task1.model;
+package by.tms.task1.model;
 
+import lombok.Getter;
+
+@Getter
 public class PersonRegistry {
 
     private Person[] persons;
@@ -10,7 +13,4 @@ public class PersonRegistry {
         this.persons = persons;
     }
 
-    public Person[] getPersons() {
-        return persons;
-    }
 }
