+++ Lesson8/src/main/java/by/tms/task1/utils/Constants.java
@@ -1,4 +1,4 @@
-package Task1.utils;
+package by.tms.task1.utils;
 
 
 public class Constants {
