+++ Lesson8/src/main/java/by/tms/task2/model/Car.java
@@ -1,0 +1,4 @@
+package by.tms.task2.model;
+
+public class Car {
+}
